requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
web3>=6.11.0
aiohttp>=3.9.0
requests>=2.31.0
eth-abi>=4.2.0
eth-utils>=2.3.0
solidity-parser>=0.1.0
py-solc-x>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
